// src/components/types/pokemon.ts
export type Ability = {
    name: string;
    slot: number;
    isHidden: boolean;
};

export type Sprites = {
    backDefault: string;
    backFemale?: string;
    backShiny?: string;
    backShinyFemale?: string;
    frontDefault: string;
    frontFemale?: string;
    frontShiny?: string;
    frontShinyFemale?: string;
};